Add PipelineVisitor for RunInference during construction time
